refactor(examples): one _common/mockserver.py instead of three drifting copies
The three _mockserver.py copies (141/102/101 lines) become 20-line re-export shims over
examples/_common/mockserver.py. A shim per directory rather than an edited import line in each of
the 18 scripts: the examples are run directly, so only their own directory is on sys.path — and
this way every existing `from _mockserver import ...` keeps working, including the sys.path hops
_mirage.py and _llamaindex.py already make into using-official-sdk/.

The three drifts are resolved rather than averaged:
  * health-check timeout: 10s, not 2 — the official-sdk copy's comment explains it (a remote
    deployment may be a trans-continental HTTPS hop); 2 was the local-only assumption.
  * the certifi/SSL_CERT_FILE block, which the BYO copy lacked entirely.
  * _url_from_argv, which only the BYO copy had. Kept, and serve_or_connect still falls back to
    it, so bring-your-own-corpus/run.py keeps honouring --url without parsing it itself — but now
    every directory gets that behaviour.

Verified: shims import from all three directories with ROOT resolving to the repo root; --url
parsing reachable from each; slack.py, gdrive.py (google_service_account_info) and gmail.py
(google_oauth_user) run end-to-end; bring-your-own-corpus/run.py validates 31 records and serves
35 docs across all 11 sources (it runs until Ctrl+C by design, so it is checked by reading its
output, not by exit code). tests/test_sdk.py + test_llamaindex.py: 17 passed. pytest 707 passed.
