+++ examples/_common/__init__.py
@@ -1,0 +1,1 @@
+"""Plumbing shared by the example directories (see ``mockserver``)."""
